+++ lib/redox/models/visit.rb
@@ -12,9 +12,16 @@ module Redox
       property :Duration, from: :duration, required: false
       property :VisitNumber, from: :visit_number, required: false
       property :AccountNumber, from: :account_number, required: false
+      property :Status, from: :status, required: false
+      property :Type, from: :type, required: false
+      property :Reason, from: :reason, required: false
+      property :Equipment, from: :equipment, required: false
+      property :CancelReason, from: :cancel_reason, required: false
+      property :DischargeDateTime, from: :end, required: false
 
       alias_method :insurances, :Insurances
       alias_method :start, :VisitDateTime
+      alias_method :end, :DischargeDateTime
 
       def department=(v)
         self[:Location] ||= DEFAULT_LOCATION
@@ -28,6 +35,12 @@ module Redox
         self
       end
 
+      def add_equipment(description: nil, code: nil)
+        self[:Equipment] ||= []
+        self[:Equipment] << { Description: description, Code: code }
+        self
+      end
+
       def insurances
         self[:Insurances] = self[:Insurances].map {|ins| ins.is_a?(Redox::Models::Insurance) ? ins : Insurance.new(ins) }
       end
@@ -35,7 +48,7 @@ module Redox
       def to_h
         result = super.to_h
 
-        %w[VisitDateTime].each do |k|
+        %w[VisitDateTime DischargeDateTime].each do |k|
           result[key][k] = Redox::Models.format_datetime(result[key][k])
         end