+++ CHANGELOG.md
@@ -4,9 +4,29 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.6.1] - 2021-03-12
+### Added
+- Media model file helper to base64 files under 200KB
+- Media model available= sets to valid Redox values with true/false
+- Medai model provider entry
+
+## [1.6.0] - 2021-02-04
+### Added
+- Media model
+- MediaUpdate Model
+- Media#create
+- Visit#CancelReason
+- Visit#DischargeDateTime
+
+## [1.5.2] - 2021-01-08
+### Added
+- Visit#Type
+- Visit#Status
+- Visit#Reason
+- Visit#Equipment
+
 ## [1.5.1] - 2020-12-22
 ### Added
-
 - Contacts model
 - Patient#contacts
 
@@ -35,12 +55,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Visit#as\_json
 
 ## [1.3.0] - 2020-09-17
-### Added 
+### Added
 - Transaction model
 - Financial#create
 
 ## [1.2.0] - 2020-07-30
-### Added 
+### Added
 - Visit model
 - Model#insurances helper for Patient.Insurances || Visit.Insurances
 
@@ -53,7 +73,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Moving to Requst classes instead of mixing in logic to models (kept backwards compatibility)
 - Added potential matches to responses and implemented for patient search
 
-=======
 ### Added
 - PotentialMatches class
 
@@ -138,6 +157,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Initial Release
 
+[1.6.1]: https://github.com/WeInfuse/redox/compare/v1.6.0...v1.6.1
+[1.6.0]: https://github.com/WeInfuse/redox/compare/v1.5.2...v1.6.0
+[1.5.2]: https://github.com/WeInfuse/redox/compare/v1.5.1...v1.5.2
 [1.5.1]: https://github.com/WeInfuse/redox/compare/v1.5.0...v1.5.1
 [1.5.0]: https://github.com/WeInfuse/redox/compare/v1.4.0...v1.5.0
 [1.4.0]: https://github.com/WeInfuse/redox/compare/v1.3.1...v1.4.0